requirements.txt
ecdsa==0.13
Fabric==1.10.2
guppy==0.1.10
paramiko==1.16.0
pycrypto==2.6.1
PyMySQL==0.7.1
python-memcached==1.57
requests==2.9.1
six==1.10.0
web.py==0.37
wheel==0.24.0
